django-advanced/requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-stdimage==5.0.2
gunicorn==20.0.4
psycopg2-binary==2.8.4
